requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.11.0=py38h578d9bd_0
aiohttp=3.7.3=py38h25fe258_0
alabaster=0.7.12=py_0
argh=0.26.2=py38_0
argon2-cffi=20.1.0=py38h25fe258_2
astor=0.8.1=pyh9f0ad1d_0
astroid=2.4.2=py38_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py_1000
async_generator=1.10=py_0
atomicwrites=1.4.0=py_0
attrs=20.2.0=py_0
autopep8=1.5.4=py_0
babel=2.8.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h7b6447c_1000
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.1=py_0
cachetools=4.1.1=py_0
certifi=2020.12.5=py38h578d9bd_1
cffi=1.14.3=py38he30daa8_0
chardet=3.0.4=py38_1003
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
cryptography=3.1.1=py38h1ba5d50_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py_2
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20200713=py_0
docutils=0.16=py38_1
entrypoints=0.3=py38_0
expat=2.2.10=he6710b0_2
flake8=3.8.4=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py38_1
gast=0.3.3=py_0
glib=2.56.2=hd408876_0
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.33.2=py38heead2fc_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=nompi_py38h7442b35_105
hdf5=1.10.6=nompi_h7c3c948_1111
icu=58.2=he6710b0_3
idna=2.10=py_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
isort=5.6.4=py_0
jedi=0.17.1=py38_0
jeepney=0.4.3=py_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=habf39ab_1
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py38_0
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
keras=2.4.3=py_0
keras-preprocessing=1.1.0=py_0
keyring=21.4.0=py38_1
kiwisolver=1.3.1=py38h82cb98a_0
krb5=1.17.2=h926e7f8_0
lazy-object-proxy=1.4.3=py38h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=h14c3975_1
libev=4.33=h516909a_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libnghttp2=1.41.0=h8cfc5f6_2
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h7b6447c_0
libspatialindex=1.9.3=he6710b0_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.3=py38h578d9bd_0
matplotlib-base=3.3.3=py38h5c7f4ab_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
multidict=4.7.5=py38h1e0a361_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
nibabel=3.2.1=pyhd8ed1ab_0
nilearn=0.7.0=pyhd8ed1ab_0
ninja=1.10.2=py38hff7bd54_0
notebook=6.1.5=py38h578d9bd_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
numpydoc=1.1.0=py_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h27cfd23_0
opt_einsum=3.3.0=py_0
packaging=20.4=py_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pathtools=0.1.2=py_1
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38he98fc37_0
pip=20.3=py38h06a4308_0
pluggy=0.13.1=py38_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py38hadf7658_1
psutil=5.7.2=py38h7b6447c_0
ptyprocess=0.6.0=py38_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pydicom=2.1.0=pyhd3deb0d_0
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pygments=2.7.1=py_0
pyjwt=2.0.0=pyhd8ed1ab_0
pylint=2.6.0=py38_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyqt5=5.15.2=pypi_0
pyqt5-sip=12.8.1=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.35.1=py_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2020.4=pyhd3eb1b0_0
pyxdg=0.26=py_0
pyyaml=5.3.1=py38h7b6447c_1
pyzmq=19.0.2=py38he6710b0_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=h5867ecd_1
qtawesome=1.0.1=py_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rope=0.18.0=py_0
rsa=4.6=pyh9f0ad1d_0
rtree=0.9.4=py38_1
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
secretstorage=3.1.2=py38_0
send2trash=1.5.0=py_0
setuptools=50.3.2=py38h06a4308_2
sip=4.19.24=py38he6710b0_0
six=1.15.0=py38h06a4308_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.2.2=py_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.5=py38_0
spyder-kernels=1.9.4=py38_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.4.0=pyhd8ed1ab_0
tensorboard-plugin-wit=1.7.0=pyh9f0ad1d_0
tensorflow=2.2.0=mkl_py38h6d3daf0_0
tensorflow-base=2.2.0=mkl_py38h5059a2d_0
tensorflow-estimator=2.2.0=pyh95af2a2_0
termcolor=1.1.0=py_2
terminado=0.9.1=py38h32f6830_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.0.4=py38h7b6447c_1
tqdm=4.56.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
ujson=4.0.1=py38he6710b0_0
urllib3=1.25.11=py_0
watchdog=0.10.3=py38_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.0=pyhd3eb1b0_0
wrapt=1.11.2=py38h7b6447c_0
wurlitzer=2.0.1=py38_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.30.0=py_0
yarl=1.6.3=py38h25fe258_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h6597ccf_2
